Path resample to given interval
Also includes “geo interval” not really sure about what that means.
Hopefully the comment will help people understand the context it should
be used in.
