+++ path.go
@@ -151,92 +151,6 @@ func (p *Path) Transform(projector Projector) *Path {
 	return p
 }
 
-// Resample converts the path into totalPoints-1 evenly spaced segments.
-func (p *Path) Resample(totalPoints int) *Path {
-	// degenerate case
-	if len(p.PointSet) <= 1 {
-		return p
-	}
-
-	if totalPoints <= 0 {
-		p.PointSet = make([]Point, 0)
-		return p
-	}
-
-	// if all the points are the same, treat as special case.
-	equal := true
-	for _, point := range p.PointSet {
-		if !p.PointSet[0].Equals(&point) {
-			equal = false
-			break
-		}
-	}
-
-	if equal {
-		if totalPoints > p.Length() {
-			// extend to be requested length
-			for p.Length() != totalPoints {
-				p.PointSet = append(p.PointSet, p.PointSet[0])
-			}
-
-			return p
-		}
-
-		// contract to be requested length
-		p.PointSet = p.PointSet[:totalPoints]
-		return p
-	}
-
-	points := make([]Point, 1, totalPoints)
-	points[0] = p.PointSet[0] // start stays the same
-
-	// first distance we're looking for
-	step := 1
-	totalDistance := 0.0
-	distance := 0.0
-	distances := make([]float64, len(p.PointSet)-1)
-	for i := 0; i < len(p.PointSet)-1; i++ {
-		distances[i] = p.PointSet[i].DistanceFrom(&p.PointSet[i+1])
-		totalDistance += distances[i]
-	}
-
-	currentDistance := totalDistance / float64(totalPoints-1)
-	currentLine := &Line{} // declare here and update has nice performance benefits
-	for i := 0; i < len(p.PointSet)-1; i++ {
-		currentLine.a = p.PointSet[i]
-		currentLine.b = p.PointSet[i+1]
-
-		currentLineDistance := distances[i]
-		nextDistance := distance + currentLineDistance
-
-		for currentDistance <= nextDistance {
-			// need to add a point
-			percent := (currentDistance - distance) / currentLineDistance
-			points = append(points, Point{
-				currentLine.a[0] + percent*(currentLine.b[0]-currentLine.a[0]),
-				currentLine.a[1] + percent*(currentLine.b[1]-currentLine.a[1]),
-			})
-
-			// move to the next distance we want
-			step++
-			currentDistance = totalDistance * float64(step) / float64(totalPoints-1)
-			if step == totalPoints-1 { // weird round off error on my machine
-				currentDistance = totalDistance
-			}
-		}
-
-		// past the current point in the original line, so move to the next one
-		distance = nextDistance
-	}
-
-	// end stays the same, to handle round off errors
-	if totalPoints != 1 { // for 1, we want the first point
-		points[totalPoints-1] = p.PointSet[len(p.PointSet)-1]
-	}
-	(&p.PointSet).SetPoints(points)
-	return p
-}
-
 // Decode is deprecated, use NewPathFromEncoding
 func Decode(encoded string, factor ...int) *Path {
 	return NewPathFromEncoding(encoded, factor...)
